Implement filter in getAll, separate config columns places and change tag logout in menu sidebar
