+++ src/app/shared/components/menu-sidebar.js
@@ -31,7 +31,7 @@ const MenuSidebar = () => {
             <div className="container-menu">
                 <Link className="nav-link" to="/auth/home"><HomeIcon /> Home</Link>
                 <Link className="nav-link" to="/auth/places"><PlaceIcon /> Locais</Link>
-                <a href="#" className="nav-link" onClick={() => onLogout()}><ExitToAppIcon /> Logout</a>
+                <span className="nav-link" onClick={() => onLogout()}><ExitToAppIcon /> Logout</span>
             </div>
         </Sidebar>
     )
